+++ Runtime/RasaNLP.cs
@@ -19,6 +19,7 @@ All rights reserved.
 
 using UnityEngine;
 using Voxell.NLP.Tokenize;
+using Voxell.NLP.SentenceDetect;
 using Voxell.Inspector;
 
 namespace Voxell.Rasa
@@ -26,13 +27,25 @@ namespace Voxell.Rasa
   [System.Serializable]
   public struct RasaNLP
   {
+    /// <summary>Checks if the tools have been initialized.</summary>
+    public bool IsInitialized => maximumEntropyTokenizer != null;
+
     public EnglishMaximumEntropyTokenizer maximumEntropyTokenizer;
-    [StreamingAssetFilePath] public string maximumEntropyTokenizerModel;
+    [SerializeField, StreamingAssetFilePath]
+    private string maximumEntropyTokenizerModel;
+
+    public EnglishMaximumEntropySentenceDetector maximumEntropySentenceDetector;
+    [SerializeField, StreamingAssetFilePath]
+    private string maximumEntropySentenceDetectorModel;
 
     public void Init()
     {
       maximumEntropyTokenizer = new EnglishMaximumEntropyTokenizer(
         FileUtil.GetStreamingAssetFilePath(maximumEntropyTokenizerModel)
+      );
+
+      maximumEntropySentenceDetector = new EnglishMaximumEntropySentenceDetector(
+        FileUtil.GetStreamingAssetFilePath(maximumEntropySentenceDetectorModel)
       );
     }
   }